Move shared code into shared utils, rename to reflect actual purpose.
